Unify coach today/week/plan tabs into one coach home
The coach section had three tabs (today, week, plan) showing overlapping
views of the same week. The home page now lays out the whole week —
completed, skipped, missed, and upcoming sessions — with today's card
expanded and actionable (log done/skip/bare-min, undo), plus inline plan
editing behind an 'edit plan' toggle. Week paging (prev/this/next) carries
over, and a malformed ?start param now falls back to the current week
instead of a 500. Tabs are now home / metrics / history / sync.
